Find Palindorme Index
Given a string of lowercase letters, determine the index of the character whose removal will make the string a palindrome. If the string is already a palindrome, then print −1−1. There will always be a valid solution.

Test Case 1(aaabaaab): Removing b at index 33 results in a palindrome, so we print 3.

Test Case 2(baabaa): Removing b at index 00 results in a palindrome, so we print 0.

Test Case 3(aaaaaa): This string is already a palindrome, so we print -1; however, 0, 1, and 2 are also all acceptable answers, as the string will still be a palindrome if any one of the characters at those indices are removed.